pwa: add ServiceWorkerProvider
