+++ pages/_app.jsx
@@ -16,6 +16,7 @@ import PipPortalProvider from '../components/common/PipPortal';
 import ErrorBoundary from '../components/core/ErrorBoundary';
 import Script from 'next/script';
 import { reportWebVitals as reportWebVitalsUtil } from '../utils/reportWebVitals';
+import ServiceWorkerProvider from '../app/components/ServiceWorkerProvider';
 
 import { Ubuntu } from 'next/font/google';
 
@@ -44,40 +45,6 @@ function MyApp(props) {
       console.error('Analytics initialization failed', err);
     });
 
-    if (process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
-      // Register PWA service worker generated via @ducanh2912/next-pwa
-      const register = async () => {
-        try {
-          const registration = await navigator.serviceWorker.register('/sw.js');
-
-          window.manualRefresh = () => registration.update();
-
-          if ('periodicSync' in registration) {
-            try {
-              const status = await navigator.permissions.query({
-                name: 'periodic-background-sync',
-              });
-              if (status.state === 'granted') {
-                await registration.periodicSync.register('content-sync', {
-                  minInterval: 24 * 60 * 60 * 1000,
-                });
-              } else {
-                registration.update();
-              }
-            } catch {
-              registration.update();
-            }
-          } else {
-            registration.update();
-          }
-        } catch (err) {
-          console.error('Service worker registration failed', err);
-        }
-      };
-      register().catch((err) => {
-        console.error('Service worker setup failed', err);
-      });
-    }
   }, []);
 
   useEffect(() => {
@@ -157,6 +124,7 @@ function MyApp(props) {
           Skip to app grid
         </a>
         <SettingsProvider>
+          <ServiceWorkerProvider />
           <PipPortalProvider>
             <div aria-live="polite" id="live-region" />
             <Component {...pageProps} />
